Provisionally cleaning up older --macro calls from the build files.
It seems keep is a macro in newer versions of haxe, and is the preferred
way anyway. I can't manage to get the compiler to remove the Main file
no matter what, so removing it provisionally. It causes FD and probably
more to choke on the code complete step when it throws errors from the
--macro option.

This can be readded if keep is really needed, but it doesn't appear to
be (or if it's using the incorrect keep syntax then that could be
corrected).