Fix openFile not working on Linux.
Thanks to LeotomasMC for some assistance with this.

I noticed and was quite confused by the fact that certain things in Flixel failed to work on Linux, such as FlxG.openURL, which returned "xdg-open: unexpected argument '&'". This appears to fix that and seems to work quite fine, with that segment of Flixel now working once more.

May be good to have someone more qualified than I check this out before er... merging, as I'm not exactly the best programmer around.