Fix rest of `WebGL2RenderContext` indentation.
This all could have been one big change, but the diff view would be very hard to read. Breaking it up like this allows Git to highlight the whitespace changes for you. Otherwise, it would show that 1245 lines were removed and replaced by 1245 new lines, not recognizing that any of those lines match.